♻️ Extract game play from console output
The end goal here is to be able to invoke game play from a variety of
contexts, whether it be a test, running a single game from a console,
or running an entire tournament. Rather than doing a bunch of `print()`s,
the `play()` function now emits events. The `console.py` file
subscribes to these events and handles all the output. The idea here is
that we can now invoke `play()` in other contexts and—as long as
nothing's imported `console`—those events won't trigger any output.