♻️ Extract winner code out of presentation
Note this had the side effect of fixing a bug that I'm not sure would
have ever manifested? The `game_ended` signal handler was returning a 1
when the game ended in a tie. This code was a holdover from when the
logic was returning a 1 error code to the console. Anyhow, it's gone.